requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
requests>=2.31.0
python-docx>=1.0.0
openpyxl>=3.1.0
